requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
PyYAML==5.4.1
requests==2.26.0
selenium==3.141.0
soupsieve==2.2.1
urllib3==1.26.6
